remove usage of size_t in arr.c and hash.c
size_t is annoying to deal with because it is not the same on all
platforms (duh), but most places where it is being used it is assumed to
be a uint32_t so make that the case.

+++ include/datastructures/arr.h
@@ -11,22 +11,22 @@
 #include <stdint.h>
 
 struct arr {
-	size_t len;
-	size_t cap;
-	size_t item_size;
+	uint32_t len;
+	uint32_t cap;
+	uint32_t item_size;
 	uint8_t *e;
 };
 
-void arr_init(struct arr *arr, size_t initial, size_t item_size);
+void arr_init(struct arr *arr, uint32_t initial, uint32_t item_size);
 void arr_destroy(struct arr *arr);
-size_t arr_push(struct arr *arr, const void *item);
+uint32_t arr_push(struct arr *arr, const void *item);
 void *arr_pop(struct arr *arr);
 void *arr_peek(struct arr *arr, uint32_t i);
-void *arr_get(const struct arr *arr, size_t i);
-void arr_del(struct arr *arr, size_t i);
+void *arr_get(const struct arr *arr, uint32_t i);
+void arr_del(struct arr *arr, uint32_t i);
 void arr_clear(struct arr *arr);
-void arr_grow_by(struct arr *arr, size_t size);
-void arr_grow_to(struct arr *arr, size_t size);
+void arr_grow_by(struct arr *arr, uint32_t size);
+void arr_grow_to(struct arr *arr, uint32_t size);
 
 typedef int32_t (*sort_func)(const void *a, const void *b, void *ctx);
 void arr_sort_range(struct arr *arr, uint32_t start, uint32_t end, void *ctx, sort_func func);
